+++ self-tests.yml
@@ -515,9 +515,6 @@ tasks:
     name: test-evergreen
   - <<: *run-go-test-suite-with-mongodb
     tags: ["db", "test"]
-    name: test-taskoutput
-  - <<: *run-go-test-suite-with-mongodb
-    tags: ["db", "test"]
     name: test-thirdparty
   - <<: *run-go-test-suite-with-mongodb
     tags: ["db", "test"]
